requirements.txt
# MCP Core Dependencies
mcp>=1.13.0
fastmcp>=2.11.0
httpx>=0.28.0
python-dotenv>=1.1.0
pydantic>=2.11.0
pydantic-settings>=2.10.0

# LangChain Ecosystem (AI/LLM Integration)
langchain>=0.3.27
langchain-core>=0.3.74
langchain-community>=0.3.27
langchain-mistralai>=0.2.11
langsmith>=0.4.15

# Mistral AI API
mistralai>=1.9.0

# Text Processing and Sentiment Analysis
textblob>=0.19.0
nltk>=3.9.0
vaderSentiment>=3.3.2

# HTTP and Async Support
requests>=2.32.0
aiofiles>=24.1.0
aiohttp>=3.12.0
httpcore>=1.0.0

# Logging and Utilities
colorlog>=6.9.0
click>=8.2.0
rich>=14.1.0
typing-extensions>=4.14.0

# Additional Dependencies
PyYAML>=6.0.0
python-dateutil>=2.9.0
tenacity>=9.1.0
packaging>=25.0.0
